feat: auto generation of the mode runner and mission executor

+++ include/ros2_uav_px4/utils/uav_cpp_ros2_conversions.hpp
@@ -34,7 +34,8 @@ uav_cpp::pipelines::PoseHeading convertFromRosMsg(const ros2_uav_interfaces::msg
 {
   uav_cpp::pipelines::PoseHeading setpoint;
   setpoint.timestamp =
-    std::chrono::nanoseconds{msg.header.stamp.sec * 1000000000 + msg.header.stamp.nanosec};
+    std::chrono::nanoseconds{static_cast<uint64_t>(msg.header.stamp.sec) * 1000000000 +
+    static_cast<uint64_t>(msg.header.stamp.nanosec)};
   setpoint.frame_id = msg.header.frame_id;
   setpoint.position = tf2::Vector3(msg.position.x, msg.position.y, msg.position.z);
   setpoint.velocity = tf2::Vector3(msg.velocity.x, msg.velocity.y, msg.velocity.z);
